Replace the current fragment labels with random labels
* src/base/construct.v:
* src/base/equal.v:
* src/base/primitive.v:
* src/base/same.v:
Replace the current fragment labels with random labels.  This is
because I seem to spend a fair amount of time thinking of labels to
suit the content of the fragments.  A manual label depends on semantic
information from the fragment content, which is liable to change and
make the label obsolete unless it too is changed.  Therefore, it seems
best to use opaque labels that are independent of the semantics of the
fragment content.

* .emacs/yasnippet/coq-mode/code-delimter: Change the yasnippet code
  to use a randomly generated label for fragments.

* .emacs/yasnippet/coq-mode/.yas-setup.el: New file.  Contains Emacs
  Lisp code used in the Coq mode snippets.